feat(packages): Refactoring package structure
