+++ cmd/httpd/internal/handlers/translations.go
@@ -2,10 +2,11 @@ package handlers
 
 import (
 	"encoding/json"
+	"net/http"
+
 	"github.com/julienschmidt/httprouter"
 	"github.com/sirupsen/logrus"
 	"mayaleng.org/engine/internal/translator"
-	"net/http"
 )
 
 type translations struct {
