+++ cmd/httpd/main.go
@@ -2,20 +2,21 @@ package main
 
 import (
 	"context"
-	"github.com/kelseyhightower/envconfig"
-	"github.com/sirupsen/logrus"
-	"mayaleng.org/engine/cmd/httpd/internal/handlers"
-	"mayaleng.org/engine/internal/platform/database"
-	"mayaleng.org/engine/internal/platform/envs"
-	"mayaleng.org/engine/version"
 	"net/http"
 	"os"
 	"os/signal"
 	"syscall"
+
+	"github.com/kelseyhightower/envconfig"
+	"github.com/sirupsen/logrus"
+	"mayaleng.org/engine/cmd/httpd/internal/handlers"
+	"mayaleng.org/engine/internal/platform/database"
+	"mayaleng.org/engine/internal/platform/types"
+	"mayaleng.org/engine/version"
 )
 
 func main() {
-	var envs envs.Envs
+	var envs types.Envs
 	envError := envconfig.Process("app", &envs)
 
 	logrus.SetFormatter(&logrus.TextFormatter{
