+++ internal/platform/linguakit/word.go
