+++ internal/platform/types/doc.go
@@ -1,0 +1,4 @@
+/*
+Package types store custom types used in the whole project
+*/
+package types
